Cleanup and hotfix for refresh playlist after add
