+++ shauncloud/Controllers/SearchDetailViewController.swift
@@ -14,48 +14,22 @@ class SearchDetailViewController: UIViewController, UITableViewDelegate, UITable
     
     @IBOutlet weak var tableView: UITableView!
     
-    var networking = Networking()
-    var playlistID = [String]()
+    private var networking = Networking()
+    private var playlistID = [String]()
+    private var tracks = [String]()
+
+    private var refreshControl: UIRefreshControl!
+
+    // MARK: View lifecycle
 
     override func viewDidLoad() {
         super.viewDidLoad()
-        // Do any additional setup after loading the view.
+        self.setUp()
     }
     
     override func viewWillAppear(animated: Bool) {
         super.viewWillAppear(animated)
-        
-        // Check to see if any playlists have been added since load.
-        networking.getPlaylist() { responseObject, error in
-            
-            
-            var newPlaylist = [String]()
-            var newArtURL = [String]()
-            
-            if let json = responseObject {
-                for index in 0..<json.count {
-                    
-                    let artURL = json[index]["artwork_url"].stringValue
-                    let title = json[index]["title"].stringValue
-                    let id = json[index]["id"].stringValue
-                    self.playlistID.append(id)
-                    newPlaylist.append(title)
-                    newArtURL.append(artURL)
-                }
-                
-                if newArtURL == Playlists.userPlaylists.playlistArtURL {
-                } else {
-                    Playlists.userPlaylists.playlistArtURL = newArtURL
-                    self.tableView.reloadData()
-                }
-                
-                if newPlaylist == Playlists.userPlaylists.playlistTitles {
-                } else {
-                    Playlists.userPlaylists.playlistTitles = newPlaylist
-                    self.tableView.reloadData()
-                }
-            }
-        }
+        self.checkForUpdatedPlaylists()
     }
 
     override func didReceiveMemoryWarning() {
@@ -63,13 +37,15 @@ class SearchDetailViewController: UIViewController, UITableViewDelegate, UITable
         // Dispose of any resources that can be recreated.
     }
     
+    // MARK: Tableview
+    
     func numberOfSectionsInTableView(tableView: UITableView) -> Int {
         return 1
     }
     
     func tableView(tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        if let count = User.currentUser.playlistCount {
-            return count
+        if Playlists.userPlaylists.playlistIDs.count != 0 {
+            return Playlists.userPlaylists.playlistIDs.count
         } else {
             return 0
         }
@@ -94,18 +70,17 @@ class SearchDetailViewController: UIViewController, UITableViewDelegate, UITable
         }
         
         return cell
-        
     }
-    
-    var tracks = [String]()
     
     func tableView(tableView: UITableView, didSelectRowAtIndexPath indexPath: NSIndexPath) {
         
         let row = indexPath.row
         let selectedPlaylist = self.playlistID[row]
         
-        self.tracks.removeAll(keepCapacity: false)
+        // Clean up track array
+        self.cleanUp()
         
+        // Find selected playlist to add track
         networking.getPlaylist() { response, error in
             if let json = response {
                 let trackCount = json[row]["tracks"].count
@@ -113,20 +88,58 @@ class SearchDetailViewController: UIViewController, UITableViewDelegate, UITable
                     let id = json[row]["tracks", index]["id"].stringValue
                     self.tracks.append(id)
                 }
+                // Insert new track into local array
                 if let newTrack = self.trackID {
                     self.tracks.insert(newTrack, atIndex: 0)
                 }
             } else {
                 print(error)
             }
-            
+            // Add new track to playlist in Soundcloud
             self.networking.addTrack(selectedPlaylist, tracks: self.tracks) { response, error in
                 self.navigationController?.popToRootViewControllerAnimated(true)
-
             }
-        
         }
-
+    }
     
+    // MARK: Helpers
+    
+    private func checkForUpdatedPlaylists() {
+        // Check to see if any playlists have been added since load.
+        networking.getPlaylist() { responseObject, error in
+            var newPlaylist = [String]()
+            var newArtURL = [String]()
+            
+            if let json = responseObject {
+                for index in 0..<json.count {
+                    let artURL = json[index]["artwork_url"].stringValue
+                    let title = json[index]["title"].stringValue
+                    let id = json[index]["id"].stringValue
+                    self.playlistID.append(id)
+                    newPlaylist.append(title)
+                    newArtURL.append(artURL)
+                }
+                Playlists.userPlaylists.playlistArtURL = newArtURL
+                Playlists.userPlaylists.playlistTitles = newPlaylist
+                self.tableView.reloadData()
+            }
+        }
+    }
+    
+    private func setUp() {
+        refreshControl = UIRefreshControl()
+        refreshControl.attributedTitle = NSAttributedString(string: "Pull to refresh")
+        refreshControl.addTarget(self, action: #selector(SearchDetailViewController.refresh(_:)), forControlEvents: UIControlEvents.ValueChanged)
+        self.tableView.addSubview(refreshControl)
+    }
+    
+    private func cleanUp() {
+        self.tracks.removeAll(keepCapacity: false)
+    }
+    
+    func refresh(sender:AnyObject) {
+        // Code to refresh table view
+        self.tableView.reloadData()
+        refreshControl.endRefreshing()
     }
 }
