+++ shauncloud/Controllers/SearchViewController.swift
@@ -12,64 +12,47 @@ class SearchViewController: UIViewController, UITableViewDelegate, UITableViewDa
 
     @IBOutlet weak var tableView: UITableView!
     
-    var array = [String]()
-    var filteredArray = [String]()
-    var filteredArrayImages = [String]()
-    var availablePlaylists = [String]()
-    var trackID = [String]()
-    var resultSearchController: UISearchController!
+    private var networking = Networking()
+
+    private var searchArray = [String]()
+    private var filteredArray = [String]()
+    private var filteredArrayImages = [String]()
+    private var availablePlaylists = [String]()
+    private var trackID = [String]()
     
-    var networking = Networking()
+    private var resultSearchController: UISearchController!
+    private let spinner: UIActivityIndicatorView = UIActivityIndicatorView  (activityIndicatorStyle: UIActivityIndicatorViewStyle.WhiteLarge)
+    private var notFoundLabel: UILabel!
     
-    let spinner: UIActivityIndicatorView = UIActivityIndicatorView  (activityIndicatorStyle: UIActivityIndicatorViewStyle.WhiteLarge)
-    var notFoundLabel: UILabel!
+    
+    // MARK: View lifecycle
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        self.navigationController?.navigationBar.hidden = true
-    
-        self.notFoundLabel = UILabel(frame: CGRectMake(0, 0, self.view.frame.width, 100))
-        
-        self.notFoundLabel.center = self.view.center
-        self.notFoundLabel.textColor = UIColor.lightGrayColor()
-        self.view.addSubview(notFoundLabel)
-        self.notFoundLabel.text = "Track not found"
-        self.notFoundLabel.textAlignment = .Center
-        self.notFoundLabel.bringSubviewToFront(self.view)
-        self.notFoundLabel.hidden = true
-        
-        spinner.color = UIColor.lightGrayColor()
-        spinner.frame = CGRectMake(0.0, 0.0, 10.0, 10.0)
-        spinner.center = self.view.center
-        self.view.addSubview(spinner)
-        spinner.bringSubviewToFront(self.view)
-        
-        self.resultSearchController = UISearchController(searchResultsController: nil)
-        self.resultSearchController.searchResultsUpdater = self
-        self.resultSearchController.dimsBackgroundDuringPresentation = false
-        self.resultSearchController.searchBar.sizeToFit()
-        self.resultSearchController?.searchBar.delegate = self
-        self.tableView.tableHeaderView = self.resultSearchController.searchBar
-        
-        
-        self.definesPresentationContext = true
+        self.setUp()
         self.tableView.reloadData()
-
-        // Do any additional setup after loading the view.
     }
     
     override func viewWillAppear(animated: Bool) {
         super.viewWillAppear(animated)
-        
+        // Update playlist in background in preparation for search detail controller
         networking.getPlaylist() { responseObject, error in
-            
-            
             if let json = responseObject {
                 for index in 0..<json.count {
                     let title = json[index]["title"].stringValue
                     self.availablePlaylists.append(title)
                 }
-            
+            }
+        }
+    }
+    
+    override func prepareForSegue(segue: UIStoryboardSegue, sender: AnyObject?) {
+        
+        if segue.identifier == "addTrack" {
+            let index = sender?.row
+            let vc = segue.destinationViewController as! SearchDetailViewController
+            if let row = index {
+                vc.trackID = self.trackID[row]
             }
         }
     }
@@ -79,6 +62,8 @@ class SearchViewController: UIViewController, UITableViewDelegate, UITableViewDa
         // Dispose of any resources that can be recreated.
     }
     
+    // MARK: Table view
+    
     func numberOfSectionsInTableView(tableView: UITableView) -> Int {
         return 1
     }
@@ -87,7 +72,7 @@ class SearchViewController: UIViewController, UITableViewDelegate, UITableViewDa
         if self.resultSearchController.active {
             return self.filteredArray.count ?? 0
         } else {
-            return self.array.count ?? 0
+            return self.searchArray.count ?? 0
         }
     }
     
@@ -96,10 +81,12 @@ class SearchViewController: UIViewController, UITableViewDelegate, UITableViewDa
         let cell = tableView.dequeueReusableCellWithIdentifier(cellIdentifier, forIndexPath: indexPath) as! TracksTableViewCell
         
         let row = indexPath.row
+        
         cell.searchCellNumberLabel?.text = String(row+1)
+        
         if self.resultSearchController.active {
-            cell.searchCellTitleLabel?.text = self.filteredArray[indexPath.row]
-            if let url = NSURL(string: self.filteredArrayImages[indexPath.row]) {
+            cell.searchCellTitleLabel?.text = self.filteredArray[row]
+            if let url = NSURL(string: self.filteredArrayImages[row]) {
                 if let data = NSData(contentsOfURL: url) {
                     cell.searchCellImage?.image = UIImage(data: data)
                 } else {
@@ -108,15 +95,13 @@ class SearchViewController: UIViewController, UITableViewDelegate, UITableViewDa
             }
             
         } else {
-            cell.searchCellTitleLabel!.text = self.array[indexPath.row]
+            cell.searchCellTitleLabel!.text = self.searchArray[row]
         }
         
+        // Setup accessory
         let addTrackIcon = UIImage(named: "addtrack")
-        
         let addTrackImage = UIImageView(image: addTrackIcon)
-        
         addTrackImage.frame = CGRect(x: 0, y: 0, width: 25, height: 25)
-        
         cell.accessoryType = .DisclosureIndicator
         cell.accessoryView = addTrackImage
         
@@ -127,20 +112,8 @@ class SearchViewController: UIViewController, UITableViewDelegate, UITableViewDa
     func tableView(tableView: UITableView, didSelectRowAtIndexPath indexPath: NSIndexPath) {
         performSegueWithIdentifier("addTrack", sender: indexPath)
     }
-        
-    override func prepareForSegue(segue: UIStoryboardSegue, sender: AnyObject?) {
-
-        if segue.identifier == "addTrack" {
-            let index = sender?.row
-            
-            let vc = segue.destinationViewController as! SearchDetailViewController
-            
-            if let row = index {
-                vc.trackID = self.trackID[row]
-            }
-            
-        }
-    }
+    
+    // MARK: Search bar delegate
     
     func searchBarSearchButtonClicked(searchBar: UISearchBar) {
         
@@ -150,11 +123,13 @@ class SearchViewController: UIViewController, UITableViewDelegate, UITableViewDa
             
             spinner.startAnimating()
             
+            // Search for query on Soundcloud
             networking.searchTracks(query) { responseObject, error in
                 if let json = responseObject {
                     self.spinner.stopAnimating()
                     self.spinner.hidesWhenStopped = true
                     
+                    // Display Not Found label if result is empty
                     if json.isEmpty {
                         self.notFoundLabel.hidden = false
                     }
@@ -175,30 +150,64 @@ class SearchViewController: UIViewController, UITableViewDelegate, UITableViewDa
     
     func searchBarCancelButtonClicked(searchBar: UISearchBar) {
         self.resultSearchController.searchBar.resignFirstResponder()
-
     }
     
     func searchBarTextDidEndEditing(searchBar: UISearchBar) {
+        // Clear search arrays
         cleanUp()
     }
   
     func updateSearchResultsForSearchController(searchController: UISearchController) {
+        // Hide Not Found label if search returns results
         self.notFoundLabel.hidden = true
-        let searchPredicate = NSPredicate(format: "SELF CONTAINS[c] %@", searchController.searchBar.text!)
         
-        let array = (self.array as NSArray).filteredArrayUsingPredicate(searchPredicate)
-        
-        self.filteredArray = array as! [String]
+        if let text = searchController.searchBar.text {
+            let searchPredicate = NSPredicate(format: "SELF CONTAINS[c] %@", text)
+            let array = (self.searchArray as NSArray).filteredArrayUsingPredicate(searchPredicate)
+            self.filteredArray = array as! [String]
+        }
         
         self.tableView.reloadData()
     }
     
-    func cleanUp() {
+    // MARK: Helpers
+    
+    private func cleanUp() {
         // Clean up arrays used to hold data
         self.filteredArray.removeAll(keepCapacity: false)
         self.filteredArrayImages.removeAll(keepCapacity: false)
         self.trackID.removeAll(keepCapacity: false)
     }
     
-
+    private func setUp() {
+        self.navigationController?.navigationBar.hidden = true
+        
+        // Instantiate Not Found Label
+        self.notFoundLabel = UILabel(frame: CGRectMake(0, 0, self.view.frame.width, 100))
+        self.notFoundLabel.center = self.view.center
+        self.notFoundLabel.textColor = UIColor.lightGrayColor()
+        self.view.addSubview(notFoundLabel)
+        self.notFoundLabel.text = "Track not found"
+        self.notFoundLabel.textAlignment = .Center
+        self.notFoundLabel.bringSubviewToFront(self.view)
+        self.notFoundLabel.hidden = true
+        
+        // Setup spinner
+        spinner.color = UIColor.lightGrayColor()
+        spinner.frame = CGRectMake(0.0, 0.0, 10.0, 10.0)
+        spinner.center = self.view.center
+        self.view.addSubview(spinner)
+        spinner.bringSubviewToFront(self.view)
+        
+        // Setup search controller
+        self.resultSearchController = UISearchController(searchResultsController: nil)
+        self.resultSearchController.searchResultsUpdater = self
+        self.resultSearchController.dimsBackgroundDuringPresentation = false
+        self.resultSearchController.searchBar.sizeToFit()
+        self.resultSearchController?.searchBar.delegate = self
+        self.tableView.tableHeaderView = self.resultSearchController.searchBar
+        
+        // Allow SearchViewController to hold after activating the search bar
+        self.definesPresentationContext = true
+    }
 }
